most tests passing. cleanup

+++ src/simplification_logic.rs
@@ -369,7 +369,7 @@ pub fn record_sample_nodes(
         if idmap[*sample as usize] != NULLTSINT {
             panic!("invalid sample list!");
         }
-        let n = tables.node(*sample);
+        let n = tables.node(*sample as usize);
         new_nodes.push(Node::<i32> {
             time: n.time,
             deme: n.deme,
